refactor(dsg): fix import
